requirements.txt
NeteaseCloudMusic
PyWinCtl
mutagen
pycryptodome
Flask
Flask-SocketIO